requirements.txt
opencv-python
numpy
pyautogui
pytesseract
pillow